requirements.txt
flask>=2.3.0
python-docx>=1.0.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
werkzeug>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
pywin32>=306; platform_system == "Windows"
# Embedding and ML dependencies
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
huggingface-hub>=0.16.0
textstat>=0.7.3
spacy>=3.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
# WebSocket and async processing
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.33.0